Add the option to fetch FortranGIS
Note that the test on TARGET is not here to protect against multiple
`find_package` calls. Multiple `find_package` calls do not matter, the
search is not repeated. The test on TARGET is not here either to
protect against multiple fetch: these would not matter either. The
test protects against a `find_package` followed by a fetch or vice-versa.